pyproject.toml
[tool.poetry]
name = "ras_spt_mod_installer"
version = "0.1.0"
description = ""
authors = ["Eszter, Gábor <mrxaxen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
wget = "^3.2"
pyinstaller = "^6.12.0"
gdown = "^5.2.0"
py7zr = "^0.22.0"
pydantic = "^2.10.6"

gitpython = "^3.1.44"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry-pyinstaller-plugin.scripts]
ras-spt-mod-installer = { source = "ras_spt_mod_installer/main.py", type = "onefile", bundle = false }

[tool.poetry.scripts]
ras-spt-mod-installer='ras_spt_mod_installer.main:main'
